==> leejeahyuk/spacetitanic02/01_numaricaldata_null_fill.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity="all"

import numpy as np
import pandas as pd

df_train = pd.read_csv('./kaggle/input/spaceship-titanic/train.csv')
df_test = pd.read_csv('./kaggle/input/spaceship-titanic/test.csv')

df_test['Transported'] = ""
df_test['testset'] = 1
df_train['testset'] = 0

df_train.columns
df_test.columns

### 0.train test 데이터 concat

df_concat = pd.concat([df_train,df_test],axis=0)

df_concat

### 1. CryoSleep==1 인 경우에 돈을 사용하지 않음

df_concat.loc[(df_concat['CryoSleep']==1) & (df_concat['RoomService'] != 0 ), 'RoomService'] = 0
df_concat.loc[(df_concat['CryoSleep']==1) & (df_concat['FoodCourt'] != 0 ), 'FoodCourt'] = 0
df_concat.loc[(df_concat['CryoSleep']==1) & (df_concat['ShoppingMall'] != 0 ), 'ShoppingMall'] = 0
df_concat.loc[(df_concat['CryoSleep']==1) & (df_concat['Spa'] != 0 ), 'Spa'] = 0
df_concat.loc[(df_concat['CryoSleep']==1) & (df_concat['VRDeck'] != 0 ), 'VRDeck'] = 0

# 없어졌는지 확인인
# df_concat.loc[df_concat['CryoSleep']==1].groupby('RoomService').count()
# df_concat.loc[df_concat['CryoSleep']==1].groupby('FoodCourt').count()
# df_concat.loc[df_concat['CryoSleep']==1].groupby('ShoppingMall').count()
# df_concat.loc[df_concat['CryoSleep']==1].groupby('Spa').count()
# df_concat.loc[df_concat['CryoSleep']==1].groupby('VRDeck').count()

### 2. Age<=12 인 경우에 돈을 사용하지 않음

df_concat.loc[(df_concat['Age']<=12) & (df_concat['RoomService'] != 0 ), 'RoomService'] = 0
df_concat.loc[(df_concat['Age']<=12) & (df_concat['FoodCourt'] != 0 ), 'FoodCourt'] = 0
df_concat.loc[(df_concat['Age']<=12) & (df_concat['ShoppingMall'] != 0 ), 'ShoppingMall'] = 0
df_concat.loc[(df_concat['Age']<=12) & (df_concat['Spa'] != 0 ), 'Spa'] = 0
df_concat.loc[(df_concat['Age']<=12) & (df_concat['VRDeck'] != 0 ), 'VRDeck'] = 0

# 없어졌는지 확인
# df_concat.loc[(df_concat['Age']<=12)].groupby('RoomService').count()
# df_concat.loc[(df_concat['Age']<=12)].groupby('FoodCourt').count()
# df_concat.loc[(df_concat['Age']<=12)].groupby('ShoppingMall').count()
# df_concat.loc[(df_concat['Age']<=12)].groupby('Spa').count()
# df_concat.loc[(df_concat['Age']<=12)].groupby('VRDeck').count()

### 3 수치형 데이터 median값으로 대체

# 나이는 61세보다 낮은 값의 중간값으로 대체
df_concat.loc[df_concat['Age'] != df_concat['Age'], 'Age'] = df_concat['Age'][df_concat['Age']<61].median()

#숫자 관련
df_concat.loc[df_concat['RoomService'].isna() == True, 'RoomService'] = df_concat['RoomService'].median()
df_concat.loc[df_concat['FoodCourt'].isna() == True, 'FoodCourt'] = df_concat['FoodCourt'].median()
df_concat.loc[df_concat['ShoppingMall'].isna() == True, 'ShoppingMall'] = df_concat['ShoppingMall'].median()
df_concat.loc[df_concat['Spa'].isna() == True, 'Spa'] = df_concat['Spa'].median()
df_concat.loc[df_concat['VRDeck'].isna() == True, 'VRDeck'] = df_concat['VRDeck'].median()

df_concat.insert(12,'TotalSpend',df_concat['RoomService']+df_concat['FoodCourt']+
df_concat['ShoppingMall']+df_concat['Spa']+df_concat['VRDeck'])

# null값 확인
df_concat.isnull().sum()

# train, test 합치고 수치형 데이터 null값으로 대체한 데이터프레임 저장
df_concat.to_csv('./kaggle/input/spaceship-titanic/df_concat.csv',index=False)
