==> leejeahyuk/spacetitanic02/02.1_data_split_totalspend.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity="all"

### 범주형 수치형 데이터 나누기
1. catboost가 범주형데이터를 잘 분석한다
    - totalspend를 그룹화 한 후 돈 관련 전부 삭제하고 catboost 돌려보기
    - 수치형 데이터를 전부 범주형으로 바꾼 후 catboost 돌려보기

import pandas as pd  
df = pd.read_csv('./kaggle/input/spaceship-titanic/concat_TotalSpendGroup.csv')

df.columns

df = df.drop(['RoomService', 'FoodCourt', 'ShoppingMall', 'Spa', 'VRDeck','TotalSpend',],axis=1)
df.columns

# Cabin 방 등급만 나누기
df['Cabin'] = df['Cabin'].str[0]

categorical = df[['PassengerId','HomePlanet','CryoSleep','Cabin','Destination','VIP','TotalSpendGroup','Transported','testset']]
numerical = df[['PassengerId','Age','testset']]

# categorical.isnull().sum()
# numerical.isnull().sum()

# df.info()

# 범주형 데이터 변환
def df_replace(dataframe):
    
    global df_train,df
    df_temp = df
    df = dataframe
    df = df.replace({'55 Cancri e':0,'PSO J318.5-22':1,'TRAPPIST-1e':2})
    df = df.replace({'Earth':0,'Europa':1,'Mars':2})
    df = df.replace({False:0,True:1})

    # Cabin : 번호  
    df['Cabin'] = df['Cabin'].str[0]
    df = df.replace({'A':0,'B':1,'C':2,'D':3,'E':4,'F':5,'G':6,'T':7})
    
    # df = pd.get_dummies(df, columns = ['Cabin'], dummy_na=False)


    # col1 = df.columns[-8:].to_list()
    # col2 = df.columns[3:-8].to_list()
    # col3 = df.columns[:3].to_list()
    # df_col = col3 + col1 + col2
    # df = df[df_col]

    # Cabin = {'A':0,'B':1,'C':2,'D':3,'E':4,'F':5,'G':6,'T':7}
    Cabin_port = {'P':0, 'S':1} #P : 배의 좌현(port side) S: 배의 우현(starboard side) 
    Destination = {'55 Cancri e': 0, 'PSO J318.5-22': 1,'TRAPPIST-1e':2}
    HomePlanet = {'Earth':0,'Europa':1,'Mars':2}
    CryoSleep = {'False':0,'True':1}
    VIP = {'False':0,'True':1}
    Transported = {'False':0,'True':1}


    df = df.astype({'HomePlanet':'float','CryoSleep':'float','Destination':'float',
    'VIP':'float',    
    })
    
# 'Cabin_A':'float','Cabin_B':'float','Cabin_C':'float','Cabin_D':'float','Cabin_E':'float','Cabin_F':'float',
#     'Cabin_G':'float','Cabin_T':'float',
# 'Transported':'float',
df_replace(df)

df_train = df

# df.info()

# from sklearn.preprocessing import StandardScaler
# from sklearn.model_selection import train_test_split
# import xgboost as xgb
# import pandas as pd
# import seaborn as sns
# from xgboost import XGBClassifier
# from sklearn.metrics import accuracy_score
# import joblib

# column_name = 'HomePlanet'
# global df_update


# df_column_name = df[df[column_name].isnull()]
# df_dropna = df.dropna(subset=[column_name])
# categorical01 = df_dropna[['PassengerId','HomePlanet','CryoSleep','Cabin','Destination','VIP','Transported','testset']]
# numerical01 = df_dropna[['PassengerId','Age','RoomService','FoodCourt','ShoppingMall','Spa','VRDeck','TotalSpend','testset']]


# X_train, X_test, y_train, y_test = train_test_split(categorical01.drop(['PassengerId',column_name,'testset'],axis=1)+numerical01.drop(['PassengerId','testset'], axis=1),categorical01[column_name], test_size=0.2, random_state=42)

# model = xgb.XGBClassifier(early_stopping_rounds=10)
# eval_set = [(X_test, y_test)]
# model.fit(X_train, y_train, eval_set=eval_set, verbose=False)#, verbose=True

# pred = model.predict(X_test)
# acc = accuracy_score(y_test,pred)
# print(acc)

df.columns

df = pd.get_dummies(df, columns = ['Cabin'], dummy_na=False)
df = pd.get_dummies(df, columns = ['TotalSpendGroup'], dummy_na=False)

df.columns

col2 = df.columns[6:-14].to_list()
col2

col1 = df.columns[-14:-6].to_list()
col2 = df.columns[6:-14].to_list()
col3 = df.columns[:6].to_list()
col4 = df.columns[-6:].to_list()

df_col = col3 + col1 +col4 + col2 
len(df.columns)
len(df_col)

df = df[df_col]
df.rename(columns={'Cabin_0.0':'Cabin_A','Cabin_1.0':'Cabin_B','Cabin_2.0':'Cabin_C','Cabin_3.0':'Cabin_D',
'Cabin_4.0':'Cabin_E','Cabin_5.0':'Cabin_F','Cabin_6.0':'Cabin_G','Cabin_7.0':'Cabin_T'}, inplace = True)
df.rename(columns={'TotalSpendGroup_0.0':'Spend_0', 'TotalSpendGroup_1000.0':'Spend_0to1k',
       'TotalSpendGroup_2500.0', 'TotalSpendGroup_5000.0',
       'TotalSpendGroup_10000.0', 'TotalSpendGroup_15000.0'}, inplace = True)

col1 = df.columns[-14:-6].to_list()
col2 = df.columns[3:-14].to_list()
col3 = df.columns[:3].to_list()
df_col = col3 + col1 + col2
df = df[df_col]
df.rename(columns={'Cabin_0.0':'Cabin_A','Cabin_1.0':'Cabin_B','Cabin_2.0':'Cabin_C','Cabin_3.0':'Cabin_D',
'Cabin_4.0':'Cabin_E','Cabin_5.0':'Cabin_F','Cabin_6.0':'Cabin_G','Cabin_7.0':'Cabin_T'}, inplace = True)

df_update.isnull().sum()
df_train = df_update

df_train.columns

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
import xgboost as xgb
import pandas as pd
import seaborn as sns
from xgboost import XGBClassifier
from sklearn.metrics import accuracy_score
import joblib

def fillnullclassifier(column_name):

    global df_update

    df_column_name = df_train[df_train[column_name].isnull()]

    df_train01 = df_train.dropna(subset=[column_name])

    X_train, X_test, y_train, y_test = train_test_split(df_train01.drop(['PassengerId',column_name,'Name'],axis=1),df_train01[column_name], test_size=0.2, random_state=42)
    model = xgb.XGBClassifier(early_stopping_rounds=10)
    eval_set = [(X_test, y_test)]
    model.fit(X_train, y_train, eval_set=eval_set, verbose=False)#, verbose=True

    # 모델 저장
    model_path = './' + 'xgb_model/' +column_name +'.model'
    # model.save_model(model_path)
    joblib.dump(model, open(model_path, 'wb'))

    pred = model.predict(X_test)
    acc = accuracy_score(y_test,pred)
    print(acc)
    
    df_column_name_null = df_column_name.drop(column_name,axis=1)

    pred = model.predict(df_column_name_null.drop(['PassengerId','Name'],axis=1))

    df_column_name[column_name] = pred

    df_update = pd.concat([df_train01,df_column_name],axis=0)

fillnullclassifier('Destination')

df_update.to_csv('./kaggle/input/spaceship-titanic/df_concat_categorical_update')

df.columns
df.isnull().sum()

df.loc[(df['testset']==0)].to_csv('./kaggle/input/spaceship-titanic/train_update.csv',index=False)
df.loc[(df['testset']==1)].to_csv('./kaggle/input/spaceship-titanic/test_update.csv',index=False)
