==> leejeahyuk/spacetitanic02/05_ensemble.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity="all"

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sbn
import numpy as np
from sklearn.preprocessing import LabelEncoder, MinMaxScaler
from sklearn.ensemble import RandomForestClassifier, BaggingClassifier, VotingClassifier, GradientBoostingClassifier, StackingClassifier
from sklearn.metrics import accuracy_score
from sklearn.model_selection import cross_val_score, GridSearchCV, KFold
from xgboost import XGBClassifier
from catboost import CatBoostClassifier

import joblib
# 모델 불러오기
cat_model_path = './' + 'model/' +'CatBoostClassifier' +'.model'
cat = joblib.load(open(cat_model_path,'rb'))
xgb_model_path = './' + 'model/' +'XGBClassifier' +'.model'
xgb = joblib.load(open(xgb_model_path,'rb'))
# xgb_model_path = './' + 'model/' +'XGBClassifier' +'.model'
# xgb_model = joblib.load(open(xgb_model_path,'rb'))

cat.get_params
xgb.get_params

# ('gb', gb), 
estimators = [('xgb', xgb), ('cat', cat)]
vt = VotingClassifier(estimators, voting='soft')

train = pd.read_csv('./kaggle/input/spaceship-titanic/train_update.csv') 
test = pd.read_csv('./kaggle/input/spaceship-titanic/test_update.csv')

from sklearn.model_selection import train_test_split
train.columns

X_train, X_test, y_train, y_test = \
train_test_split(train.drop(['PassengerId','Transported','Name','testset'],axis=1),train['Transported'], test_size=0.2, random_state=42)

# test_feature = test.drop(['PassengerId','Transported','Name','testset'],axis=1)
# test_target = test['Transported']


result_x = vt.fit(X_train, y_train)

#Accuracy
y_pred = result_x.predict(X_test)
accuracy_score(y_test, y_pred)
